requirements.txt
alembic==1.5.6
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
Flask==1.1.2
flask-restplus==0.13.0
importlib-metadata==3.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
psycopg2==2.7.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
Werkzeug==0.16.1
xlrd==1.2.0
zipp==3.4.1
